requirements.txt
fastapi == 0.115.13
uvicorn == 0.34.3
python-dotenv == 1.1.1
google-generativeai == 0.8.5
notion-client == 2.4.0
feedparser == 6.0.11